% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatWingAll.R
\name{BatWingAll}
\alias{BatWingAll}
\title{LSA + Dynamics measures for Bat wings.}
\usage{
BatWingAll(x, scale = F)
}
\arguments{
\item{x}{is the input matrix. This matrix is very specific, check the example.}

\item{scale}{if you want to transform milimeters and grames to meters and kilograms
for the  LSA and Dynamics measures calculate.}
}
\value{
The function return four LSA calculations and four Dynamics measures.

The function returns a S3 object with two list objects. First one
is the LSA and Dynamics measures, and second one is the Statistics.
}
\description{
Calculate the LSA and Dynamics measures for a Bat wing using different wings measures.
}
\examples{
library(BatWing)
data(wing)

BatWingAll(x = wing, scale=FALSE)

}
\author{
Leon-Alvarado Omar Daniel.
}
\seealso{
\url{https://github.com/oleon12/BatWing/tree/master/README.Rmd}
}
