% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatWingArea.R
\name{BatWingArea}
\alias{BatWingArea}
\title{Lifting Surface Area (LSA) for Bat wings.}
\usage{
BatWingArea(x, scale = FALSE, method = c("BM", "A", "SS", "P"))
}
\arguments{
\item{x}{is the input matrix. This matrix is very specific, check the example.}

\item{scale}{if you want to transform milimeters and grames to meters and kilograms.}

\item{method}{is the LSA calculation method ("BM","A","SS","P").}
}
\value{
The function return four LSA calculations.

The function returns a S3 object with two matrices. First one
is the LSA calculation, and second one have basic statistics for the input
variables and the LSA values.
}
\description{
Calculate the LSA for Bat wings using different wings measures.
}
\examples{
library(BatWing)
data(wing)

BatWingArea(x = wing, scale = FALSE)

}
\author{
Leon-Alvarado Omar Daniel.
}
\seealso{
\url{https://github.com/oleon12/BatWing/tree/master/README.Rmd}
}

