% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatWingDynamics.R
\name{BatWingDynamics}
\alias{BatWingDynamics}
\title{Wing's dynamic measures.}
\usage{
BatWingDynamics(x, scale = FALSE, method = c("AR", "WL", "RWL", "MPS"))
}
\arguments{
\item{x}{is the input matrix. This matrix is very specific, check the example.}

\item{scale}{if you want to transform milimeters and grames to meters and kilograms.}

\item{method}{is the LSA calculation method ("AR","WL","RWL","MPS").}
}
\value{
The function return four Dynamics meaures.

The function returns a S3 object with two matrices. First one
is the Dynamics measures, and second one have basic statistics for the input
variables and the Dynamic measures values.
}
\description{
Calculate the Dynamics measures for a Bat wing.
}
\examples{
library(BatWing)
data(dynamics)

BatWingDynamics(x = wing, scale = FALSE)

}
\author{
Leon-Alvarado Omar Daniel.
}
\seealso{
\url{https://github.com/oleon12/BatWing/tree/master/README.Rmd}
}
