% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{data}
\name{dynamics}
\alias{dynamics}
\title{Bat Wing Variables for Dynamics Measures}
\format{A data.frame object with four wing variables (columns) and ten species (rows)}
\usage{
dynamics
}
\description{
A data frame with ten species and four wing variables.
}
\author{
Leon-Alvarado, Omar Daniel
}
\keyword{datasets}
