% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\docType{data}
\name{dynamics.Rdata}
\alias{dynamics}
\alias{dynamics.Rdata}
\title{Data for Dynamics function}
\format{An object of class \code{"data.frame"}}
\usage{
data(dynamics)
}
\examples{
data(dynamics)

}
\keyword{datasets}

