% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{data}
\name{wing}
\alias{wing}
\title{Bat Wing Measures}
\format{A data.frame object with six variables (columns) and ten species (rows)}
\usage{
wing
}
\description{
A data frame with ten species and six wing variables.
}
\author{
Leon-Alvarado, Omar Daniel
}
\keyword{datasets}
