% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics.R
\docType{data}
\name{dynamics}
\alias{dynamics}
\title{Bat Wing variables for Dynamics meaures.}
\format{A data.frame object with four variables (columns) and ten species (rows)}
\usage{
dynamics
}
\description{
A data frame with four wing variables and ten species.
}
\author{
Leon-Alvarado, Omar Daniel
}
\keyword{datasets}

