% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wing.R
\docType{data}
\name{wing}
\alias{wing}
\title{Bat Wing variables for LSA calculation.}
\format{A data.frame object with six variables (columns) and tens species (rows)}
\usage{
wing
}
\description{
A data frame with six variables for LSA calculation.
}
\author{
Leon-Alvarado, Omar Daniel
}
\keyword{datasets}

